REFACTOR: moving things around to be more transparent
