+++ unet3d/scripts/script_utils.py
@@ -1,3 +1,7 @@
+import monai.losses
+import torch
+
+from unet3d.utils.pytorch import losses
 from unet3d.utils.utils import load_json
 from unet3d.models.build import build_or_load_model
 
@@ -30,3 +34,26 @@ def get_machine_config(namespace):
 def build_or_load_model_from_config(config, model_filename, n_gpus, strict=False):
     return build_or_load_model(config["model"].pop("name"), model_filename, n_gpus=n_gpus, **config["model"],
                                strict=strict)
+
+
+def load_criterion_from_config(config, n_gpus):
+    return load_criterion(config['loss'].pop("name"), n_gpus=n_gpus, loss_kwargs=config["loss"])
+
+
+def load_criterion(criterion_name, n_gpus=0, loss_kwargs=None):
+    if loss_kwargs is None:
+        loss_kwargs = dict()
+    try:
+        criterion = getattr(losses, criterion_name)(**loss_kwargs)
+    except AttributeError:
+        try:
+            criterion = getattr(torch.nn, criterion_name)(**loss_kwargs)
+        except AttributeError:
+            criterion = getattr(monai.losses, criterion_name)(**loss_kwargs)
+    if n_gpus > 0:
+        criterion.cuda()
+    return criterion
+
+
+def build_optimizer(optimizer_name, model_parameters, **kwargs):
+    return getattr(torch.optim, optimizer_name)(params=model_parameters, **kwargs)
